I've changed the way labels are created in a tree. I've implemented the decoding method.
I think the decoding method can be optimized. Note that in each iteration, we use `pop` method for l1 list. I think we can replace it with more efficient solution. Additionally, we can create a setter and getter for `neighbours` attribute of nodes, because when we modify the attribute of first object, we immediately need to alter the same attribute of second object of the same class.